actions: Replace deprecated set-output command (#26800)
GitHub [is deprecating][1] the `::set-output` command in favor of writing
key-value pairs to a file.

Also of note is that the file uses a different mechanism for setting
values containing newlines. We only had one instance of that.

[1]: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3244078234